infra (category-registration): removing SuperEntity abstraction
