+++ src/main/java/com/mercadolibrewannabe/model/validation/UniqueValidator.java
@@ -1,13 +1,12 @@
 package com.mercadolibrewannabe.model.validation;
 
-import com.mercadolibrewannabe.model.SuperEntity;
 import com.mercadolibrewannabe.model.form.AbstractForm;
 import org.springframework.validation.Errors;
 import org.springframework.validation.Validator;
 
 import java.util.Optional;
 
-public abstract class UniqueValidator<F extends AbstractForm, E extends SuperEntity> implements Validator {
+public abstract class UniqueValidator<F extends AbstractForm, E> implements Validator {
 
 	protected abstract Optional<E> findEntityByField(F form);
